+++ Problems.Tests/Problem0014Tests.cs
@@ -1,8 +1,8 @@
 ﻿namespace Problems.Tests;
 
-public class Problem0014Tests : EulerProblemTestBase<Problem0014, int>
+public class Problem0014Tests : EulerProblemTestBase<Problem0014, long>
 {
-    protected override int ExampleResult => 10;
+    protected override long ExampleResult => 10;
 
-    protected override int ProblemResult => 0;
+    protected override long ProblemResult => 0;
 }
